% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefun-package.R
\docType{package}
\name{rarefun-package}
\alias{rarefun}
\alias{rarefun-package}
\title{rarefun: Functions for Rare Events Analysis}
\description{
A collection of functions for detecting and analyzing rare events in datasets. Includes implementations of various anomaly detection methods such as isolation forest, DBSCAN clustering, and residual-based approaches. Provides tools for spatial point matching, time series extraction, and statistical testing for rare event patterns.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vlyubchich/rarefun}
  \item Report bugs at \url{https://github.com/vlyubchich/rarefun/issues}
}

}
\author{
\strong{Maintainer}: Vyacheslav Lyubchich \email{lyubchich@umces.edu} (\href{https://orcid.org/0000-0001-7936-4285}{ORCID})

Authors:
\itemize{
  \item Geneviève Nesslage (\href{https://orcid.org/0000-0003-1770-6803}{ORCID})
}

}
\keyword{internal}
