% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefun-package.R
\docType{package}
\name{rarefun-package}
\alias{rarefun}
\alias{rarefun-package}
\title{rarefun: Functions for Rare Events Analysis}
\description{
Functions for detecting and analyzing rare events in data. Includes isolation forest and DBSCAN clustering for time series residuals. Handles time series with STL/LOESS, Fourier smoothing, and estimates seasonality shifts. Also provides tools for spatial point matching, and statistical testing for rare event patterns.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vlyubchich/rarefun}
  \item Report bugs at \url{https://github.com/vlyubchich/rarefun/issues}
}

}
\author{
\strong{Maintainer}: Vyacheslav Lyubchich \email{lyubchich@umces.edu} (\href{https://orcid.org/0000-0001-7936-4285}{ORCID})

Authors:
\itemize{
  \item Geneviève Nesslage (\href{https://orcid.org/0000-0003-1770-6803}{ORCID})
}

}
\keyword{internal}
